requirements.txt
# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
motor==3.3.1
pymongo==4.6.0

# Data Pipeline
apache-airflow==2.7.1
pandas==2.1.2
numpy==1.26.1
requests==2.31.0

# Google Cloud
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.36.4

# ML/AI
tensorflow==2.14.0
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Vector Search
pymongo-vector-search==0.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
